Fix reset network with cleanup
The feature "reset network with cleanup" when used was re-creating the VR without the DHCP entries configured. Therefore, the VMs could loose their IP, and would not be able to retrieve a new lease. There was aworkaroud though; if you stopped, and then restarted a VM, the DHCP entries would be populated. However, that is not an acceptable solution; we fixed the root cause of the problem, and now the "reset network with cleanup" in the VPC tab works just fine.